Set up the initial shape of the backend
The idea is to have nginx as a proxy for Express (it will come in handy when I configure SSL and possibly to serve static files quicker). Docker-compose orchestrates firing up containers for the server, Node and MongoDB. There's also a volume container to support DB persistence and reuse between containers. It's all duct tape and fooling around basically, and a chance it may stay that way.